# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Maison de FLOF API
  version: 1.0.0
  description: Source-inventoried OpenAPI contract for every Next.js API route.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security: []
paths:
...
  /api/admin/users:
    get:
      operationId: getAdminUsers
      summary: GET /api/admin/users
      description: Source-inventoried API operation.
      x-contract-tier: stable-minimal
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GenericObject"
                  - $ref: "#/components/schemas/GenericArray"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - sessionCookie: []
    post:
      operationId: postAdminUsers
      summary: POST /api/admin/users
      description: Source-inventoried API operation.
      x-contract-tier: detailed
      responses:
        "200":
          description: Successful response
          content: &a13
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GenericObject"
                  - $ref: "#/components/schemas/GenericArray"
        "201":
          description: Resource created
          content: *a13
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
    patch:
      operationId: patchAdminUsers
      summary: PATCH /api/admin/users
      description: Source-inventoried API operation.
      x-contract-tier: detailed
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GenericObject"
                  - $ref: "#/components/schemas/GenericArray"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
    delete:
      operationId: deleteAdminUsers
      summary: DELETE /api/admin/users
      description: Source-inventoried API operation.
      x-contract-tier: detailed
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - sessionCookie: []
...
  /api/orders:
    get:
      operationId: listOrders
      summary: List orders visible to the authenticated user
      security:
        - sessionCookie: []
      parameters:
        - name: email
          in: query
          schema:
            type: string
            format: email
      responses:
        "200":
          description: Visible orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createOrder
      summary: Create an idempotent checkout order
      security:
        - sessionCookie: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 16
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Checkout"
      responses:
        "200":
          description: Existing idempotent order
        "201":
          description: Order created
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/vnpay/ipn:
    get:
      operationId: getVnpayIpn
      summary: GET /api/vnpay/ipn
      description: Simulated VNPay operation retained for demo compatibility and excluded from
        production-readiness claims.
      x-contract-tier: simulated
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GenericObject"
                  - $ref: "#/components/schemas/GenericArray"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/vnpay/return:
    get:
      operationId: getVnpayReturn
      summary: GET /api/vnpay/return
      description: Simulated VNPay operation retained for demo compatibility and excluded from
        production-readiness claims.
      x-contract-tier: simulated
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GenericObject"
                  - $ref: "#/components/schemas/GenericArray"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    Conflict:
      description: Resource state conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiError:
      type: object
      required:
        - error
        - requestId
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - TOO_MANY_REQUESTS
                - RATE_LIMITED
                - SERVICE_UNAVAILABLE
                - INTERNAL_ERROR
            message:
              type: string
            details: {}
        requestId:
          type: string
...
    Order:
      type: object
      required:
        - id
        - total
        - status
      properties:
        id:
          type: string
        total:
          type: number
        status:
          type: string
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        structuredItems:
          type: array
          items:
            type: object
            additionalProperties: true
    Checkout:
      type: object
      required:
        - items
        - paymentMethod
        - shipping
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - paintId
              - quantity
            properties:
              paintId:
                type: string
              colorId:
                type: string
              quantity:
                type: integer
                minimum: 1
        paymentMethod:
          type: string
          enum:
            - COD
            - TRANSFER
            - VNPAY
        couponCode:
          type: string
        note:
          type: string
        shipping:
          type: object
          required:
            - fullName
            - phone
            - addressLine1
            - district
            - province
          properties:
            fullName:
              type: string
            phone:
              type: string
            addressLine1:
              type: string
            addressLine2:
              type: string
            district:
              type: string
            province:
              type: string
    GenericObject:
      type: object
      additionalProperties: true
    GenericArray:
      type: array
      items:
        $ref: "#/components/schemas/GenericObject"
    Success:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          const: true
...